requirements.txt
sentence-transformers==3.0.1
torch>=2.0.0
transformers>=4.41.0
chromadb>=0.5.0
pillow>=10.0.0
pypdf>=4.2.0
tqdm>=4.66.0
numpy>=1.24.0
chroma-hnswlib==0.7.3